backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
SQLAlchemy>=2.0,<3.0
alembic>=1.13,<2.0
psycopg[binary]>=3.1,<4.0
pydantic-settings>=2.2,<3.0
email-validator>=2.2,<3.0
